+++ src/internal/Guards.ts
@@ -1,5 +1,0 @@
-export interface Guards {
-  __setText?: string | typeof NaN;
-  __setHTML?: string | typeof NaN;
-  __setAttr?: Record<string, unknown>;
-}