+++ test/readme.ts
@@ -1,15 +1,14 @@
 import {
   CustomEventElement,
-  emit,
-  nextFrame,
   qsr,
   reconcile,
-  send,
+  requestAnimationFrameOnce,
   setText,
   setValue,
-} from '../src/index.js';
+  TypedCustomEvent,
+} from '../src/exdom.js';
 
-// Type-safe custom DOM events
+// Type-safe custom DOM event details
 
 interface TodoAppEvents {
   addTodoItem: string;
@@ -47,6 +46,19 @@ function TodoApp(el: CustomEventElement<TodoAppEvents>) {
     },
   ];
 
+  // Idempotent update
+
+  function update() {
+    // Reconcile todo item components with item list
+    reconcile({
+      container: qsr(el, '.items'),
+      items,
+      create: () => TodoItem(document.createElement('li')),
+      update: (el, item) =>
+        el.dispatchEvent(new TypedCustomEvent('todoItem', { detail: item })),
+    });
+  }
+
   // Action handlers
 
   el.addEventListener('addTodoItem', (e) => {
@@ -58,7 +70,7 @@ function TodoApp(el: CustomEventElement<TodoAppEvents>) {
 
     items = [...items, newItem];
 
-    nextFrame(update);
+    requestAnimationFrameOnce(update);
   });
 
   el.addEventListener('toggleTodoItem', (e) => {
@@ -66,7 +78,7 @@ function TodoApp(el: CustomEventElement<TodoAppEvents>) {
       item.id === e.detail ? { ...item, done: !item.done } : item,
     );
 
-    nextFrame(update);
+    requestAnimationFrameOnce(update);
   });
 
   // UI Events
@@ -74,21 +86,14 @@ function TodoApp(el: CustomEventElement<TodoAppEvents>) {
   const labelInput = qsr<HTMLInputElement>(el, '[name=label]');
 
   qsr(el, '.add').addEventListener('click', () => {
-    emit<TodoAppEvents>(el, 'addTodoItem', labelInput.value);
+    el.dispatchEvent(
+      new TypedCustomEvent('addTodoItem', {
+        detail: labelInput.value,
+        bubbles: true,
+      }),
+    );
     setValue(labelInput, '');
   });
-
-  // Idempotent update
-
-  function update() {
-    // Reconcile todo item components with item list
-    reconcile({
-      container: qsr(el, '.items'),
-      items: items,
-      create: () => TodoItem(document.createElement('li')),
-      update: (el, item) => send<TodoAppEvents>(el, 'todoItem', item),
-    });
-  }
 
   update(); // Initial update
 
@@ -103,20 +108,25 @@ function TodoItem(el: CustomEventElement<TodoAppEvents>) {
 
   let data: TodoItemData;
 
-  el.addEventListener('todoItem', (e) => {
-    data = e.detail;
-    nextFrame(update);
-  });
-
-  qsr(el, '[type=checkbox]').addEventListener('click', () => {
-    // Emit action event (handled somewhere up the DOM)
-    emit<TodoAppEvents>(el, 'toggleTodoItem', data.id);
-  });
-
   function update() {
     setText(qsr(el, 'label'), data.label);
     setValue(qsr(el, '[type=checkbox]'), data.done);
   }
+
+  el.addEventListener('todoItem', (e) => {
+    data = e.detail;
+    requestAnimationFrameOnce(update);
+  });
+
+  qsr(el, '[type=checkbox]').addEventListener('click', () => {
+    // Emit action event (handled somewhere up the DOM)
+    el.dispatchEvent(
+      new TypedCustomEvent('toggleTodoItem', {
+        detail: data.id,
+        bubbles: true,
+      }),
+    );
+  });
 
   return el;
 }