+++ CHANGELOG.md
@@ -1,18 +1,33 @@
 # Changelog
 
+## NEXT
+
+- BREAKING: Rename `nextFrame` to `requestAnimationFrameOnce`
+- BREAKING: Remove `key: string` in `reconcile` (must be function, if any)
+- BREAKING: Remove `emit`, `send` and `relay`as they cannot be implemented in a
+  type-safe way
+- BREAKING: Remove `getValue` as it did not provide non-trivial work
+- Fix `requestAnimationFrameOnce` swallowing errors
+- Add `TypedCustomEvent` constructor and typed `dispatchEvent` signature to
+  `CustomEventTarget` (can infer details based on event target and type)
+- Use indexes as key fallback in `reconcile`
+- Handles boolean values correctly in `setAttr`
+- Migrate to one file library
+- Update dependencies
+
 ## 1.0.2
 
 - Update README
 - Update docs
 - Add tests
-- Update deps
+- Update dependencies
 
 ## 1.0.1
 
 - Improve `nextFrame` behavior
 - Fix `reconcile` key function
 - Update docs
-- Update deps
+- Update dependencies
 
 ## 1.0.0